tuned hotpath, moved tests
