+++ exposed-core/build.gradle.kts
@@ -1,4 +1,3 @@
-import org.gradle.api.tasks.testing.Test
 import org.gradle.kotlin.dsl.withType
 import org.jetbrains.kotlin.gradle.dsl.JvmTarget
 import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
@@ -28,14 +27,6 @@ dependencies {
     api(libs.kotlinx.coroutines)
     api(libs.kotlinx.jvm.datetime)
     api(libs.slf4j)
-
-    testImplementation(kotlin("test-junit5"))
-    testImplementation(libs.junit5)
-    testRuntimeOnly(libs.junit.platform.launcher)
-}
-
-tasks.withType<Test>().configureEach {
-    useJUnitPlatform()
 }
 
 tasks.withType<KotlinCompile>().configureEach {
