+++ exposed-core/src/main/kotlin/org/jetbrains/exposed/v1/core/ResultRow.kt
@@ -202,42 +202,52 @@ class ResultRow(
      * field-index slot is observed through two different column-type views — e.g. a plain `Column<Int>`
      * and the wrapping `Column<EntityID<Int>>` that share the same table+name (and are therefore
      * `equals()`-equal) but carry different [IColumnType] instances and produce different converted values.
+     *
+     * The backing arrays are allocated lazily on the first value read: rows that are constructed but never
+     * read through value-conversion — e.g. write-only DAO rows, or rows discarded before any column is
+     * accessed — allocate nothing. [cacheType] must be tracked for every populated slot (it validates the
+     * column-type view on every cache hit, not just on conflicts), so it shares [cacheData]'s lifecycle.
      */
     private class ResultRowCache(private val fieldIndex: Map<Expression<*>, Int>) {
         companion object {
             // Sentinel distinguishing "slot not yet populated" from a legitimately cached null.
-            private val UNCACHED = object {}
+            private val UNCACHED = Any()
         }
 
-        // Primary cache: indexed by fieldIndex position.
-        private val cacheData = Array<Any?>(fieldIndex.size) { UNCACHED }
-        // Column type stored alongside each primary-cache slot for type-conflict detection.
-        private val cacheType = arrayOfNulls<IColumnType<*>>(fieldIndex.size)
+        // Primary cache, indexed by fieldIndex position. Lazily allocated on the first read.
+        private var cacheData: Array<Any?>? = null
+
+        // Column type stored alongside each primary-cache slot for type-conflict detection. Shares
+        // cacheData's lazy lifecycle: non-null whenever cacheData is non-null.
+        private var cacheType: Array<IColumnType<*>?>? = null
+
         // Overflow: only allocated when a type-view conflict is encountered (rare in practice).
         private var overflow: HashMap<Pair<Expression<*>, IColumnType<*>?>, Any?>? = null
 
         /**
          * Returns the cached value for [expression], computing and storing it via [initializer] on a miss.
          *
-         * Hot path (common columns): single array-bounds check + array read — no object allocation.
-         * Overflow path (type-view conflict or expression absent from fieldIndex): falls back to a
-         * lazily-created HashMap with a [Pair] key, preserving the original semantics.
+         * Hot path (common columns): single array-bounds check + array read — no object allocation once
+         * the backing arrays exist. Overflow path (type-view conflict or expression absent from fieldIndex):
+         * falls back to a lazily-created HashMap with a [Pair] key, preserving the original semantics.
          */
         fun <T> cached(expression: Expression<*>, initializer: () -> T): T {
             val colType = (expression as? Column<*>)?.columnType
             val index = fieldIndex[expression]
 
             if (index != null) {
-                val current = cacheData[index]
+                val data = cacheData ?: Array<Any?>(fieldIndex.size) { UNCACHED }.also { cacheData = it }
+                val current = data[index]
                 when {
                     current === UNCACHED -> {
                         // First access for this slot: populate the primary cache.
                         val value = initializer()
-                        cacheData[index] = value
-                        cacheType[index] = colType
+                        data[index] = value
+                        val types = cacheType ?: arrayOfNulls<IColumnType<*>>(fieldIndex.size).also { cacheType = it }
+                        types[index] = colType
                         return value
                     }
-                    cacheType[index] == colType -> {
+                    cacheType?.get(index) == colType -> {
                         // Cache hit: same slot, same column-type view.
                         @Suppress("UNCHECKED_CAST")
                         return current as T
@@ -267,11 +277,9 @@ class ResultRow(
         fun remove(expression: Expression<*>) {
             val index = fieldIndex[expression]
             if (index != null) {
-                cacheData[index] = UNCACHED
+                cacheData?.set(index, UNCACHED)
             }
-            if (overflow != null) {
-                overflow!!.remove(expression to (expression as? Column<*>)?.columnType)
-            }
+            overflow?.remove(expression to (expression as? Column<*>)?.columnType)
         }
     }
 }
